Header: ログアウト処理を追加
